update main tests to run in a node env - tests relying on jsdom were moved to tests/renderer/*. #216.
