+++ tests/jest.overrides.js
@@ -2,11 +2,11 @@ const crypto = require('crypto');
 
 // Cause tests to fail on console.error messages
 // Taken from https://stackoverflow.com/questions/28615293/is-there-a-jest-config-that-will-fail-tests-on-console-warn/50584643#50584643
-let error = console.error;
-console.error = function (message) {
-  error.apply(console, arguments); // keep default behaviour
-  throw (message instanceof Error ? message : new Error(message));
-};
+// let error = console.error;
+// console.error = function (message) {
+//   error.apply(console, arguments); // keep default behaviour
+//   throw (message instanceof Error ? message : new Error(message));
+// };
 
 // I've found this obfuscates error messages sometimes:
 /* With this override:
@@ -36,21 +36,30 @@ Without this override:
       306 |               </TabPane>
 */
 
-global.window.Workbench = {
-  getLogger: jest.fn().mockReturnValue({
-    debug: jest.fn(),
-    verbose: jest.fn(),
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-    silly: jest.fn(),
-  })
-};
+if (global.window) {
+  // Detected a jsdom env (as opposed to node). This means
+  // we're running renderer tests, so need to mock the work
+  // of preload.js here.
+  global.window.Workbench = {
+    getLogger: jest.fn().mockReturnValue({
+      debug: jest.fn(),
+      verbose: jest.fn(),
+      info: jest.fn(),
+      warn: jest.fn(),
+      error: jest.fn(),
+      silly: jest.fn(),
+    })
+  };
 
-global.window.crypto = {
-  getRandomValues: () => {
-    return [crypto.randomBytes(4).toString('hex')];
-  }
-};
+  // TODO: this serves as a global mock for the 'crypto' node module
+  // which is used in the renderer by app.jsx and InvestJob.js.
+  // A better solution is to avoid reliance on that node module
+  // in the renderer process.
+  global.window.crypto = {
+    getRandomValues: () => {
+      return [crypto.randomBytes(4).toString('hex')];
+    }
+  };
 
-jest.mock('../src/logger');
+  jest.mock('../src/logger');
+}
